+++ backend/src/main/java/com/neon/sve/dto/serie/DatosListadoSerie.java
@@ -1,5 +1,6 @@
 package com.neon.sve.dto.serie;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import com.neon.sve.model.stock.SerieProducto;
@@ -9,7 +10,7 @@ public record DatosListadoSerie(
         Long id,
         String codigoLote,
         int cantidad,
-        double precioUni,
+        BigDecimal precioUni,
         String numeroSerie,
         String estado,
         Timestamp fechaRegistro) {
